runpod/image/requirements.txt
runpod==1.7.9
diffusers==0.32.2
transformers==4.48.3
accelerate==1.3.0
safetensors==0.5.3
Pillow==11.1.0
xformers==0.0.27
omegaconf==2.3.0
invisible-watermark==0.2.0
